backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymupdf==1.24.14
openai
httpx==0.26.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
python-multipart==0.0.6
lxml
pyalex
semanticscholar